+++ examples/first_node.py
@@ -4,8 +4,8 @@ import sys
 
 from kademlia.network import Server
 
-
 def main():
+
 	handler = logging.StreamHandler()
 	formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
 	handler.setFormatter(formatter)