+++ tests/test_config.py
@@ -1,10 +1,11 @@
-from kademlia.config import Config
+from kademlia.config import CONFIG
 
+
+# pylint: disable=too-few-public-methods
 class TestConfig:
 	# pylint: disable=no-self-use
 	def test_can_load_config(self):
-		cfg = Config()
 
-		assert cfg.alpha == 3
-		assert cfg.ksize == 20
-		assert cfg.interface == "0.0.0.0"
+		assert CONFIG.alpha == 3
+		assert CONFIG.ksize == 20
+		assert CONFIG.interface == "0.0.0.0"