+++ CHANGELOG.md
@@ -14,3 +14,5 @@
 - [chore] add docstrings, typings, and tests to all modules
 - [enhancement] add Docker support
 - [enhancement] continue support for CI (Docker, travis)
+- [enhancement] create multi-peer working examples (examples/network.py)
+- [chore] cleanup docstrings and types for all modules